Error handling in kv_manager
Use custom errors in kv manager

Organize kv_manager in files

Add some docs
