+++ src/gg20/mnemonic/error.rs
@@ -26,7 +26,7 @@ pub(super) mod mnemonic {
         #[error("File IO error: {0}")]
         FileIoErr(#[from] super::file_io::FileIoError),
         #[error("KvStore error: {0}")]
-        KvErr(#[from] crate::kv_manager::KvError),
+        KvErr(#[from] crate::kv_manager::error::KvError),
         #[error("Bip39 error {0}")]
         Bip39Error(#[from] super::bip39::Bip39Error),
         #[error("Failed to convert to SecretRecoveryKey")]
